Add A Ramped Platform
With this commit, there is now a platform that has pickups on it. There are ramps that the player can use to get on to the platform. This commit also puts a pickup on one of the bigger cube pickups; the player can jump onto the cube to get it.